# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/tasks": {
            "get": {
                "description": "Get tasks",
                "tags": [
                    "Task"
                ],
                "summary": "Get Tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.GetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Submit tasks\nPlease send the answer accordingly",
                "tags": [
                    "Task"
                ],
                "summary": "Submit Tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.SubmitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/users": {
            "get": {
                "description": "Get user detail by its bearer token",
                "tags": [
                    "User"
                ],
                "summary": "Get Auth User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.GetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update user",
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.GetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "constant.TaskType": {
            "type": "string",
            "enum": [
                "LEARNING",
                "MULTIPLE_CHOICE",
                "ESSAY",
                "MATCHING",
                "DETECTIVE",
                "LEVEL"
            ],
            "x-enum-varnames": [
                "Learning",
                "MultipleChoice",
                "Essay",
                "Matching",
                "Detective",
                "Level"
            ]
        },
        "src_entity_v1_http_task.Level": {
            "type": "object",
            "properties": {
                "qna": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/src_entity_v1_http_task.QnaLevel"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "src_entity_v1_http_task.LevelPayload": {
            "type": "object",
            "properties": {
                "level_1": {
                    "$ref": "#/definitions/src_entity_v1_http_task.Level"
                },
                "level_2": {
                    "$ref": "#/definitions/src_entity_v1_http_task.Level"
                }
            }
        },
        "src_entity_v1_http_task.QnAPayload": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "src_entity_v1_http_task.QnaLevel": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "task.GetResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "levels": {
                    "$ref": "#/definitions/src_entity_v1_http_task.LevelPayload"
                },
                "matches": {
                    "$ref": "#/definitions/task.Matches"
                },
                "point": {
                    "type": "integer"
                },
                "qnas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/src_entity_v1_http_task.QnAPayload"
                    }
                },
                "sub_task": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/constant.TaskType"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
        "task.Matches": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "task.SubmitRequest": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "id": {
                    "type": "string"
                },
                "sub_task_id": {
                    "type": "string"
                }
            }
        },
        "user.GetResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "total_point": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}